+++ lib/mini_kraken/core/goal_arg.rb
@@ -1,10 +1,12 @@
 # frozen_string_literal: true
 
+require_relative 'base_arg'
+
 module MiniKraken
   module Core
-    # The generalization of any iem that can be
-    # passed as arugement to a goal.
-    class GoalArg
+    # The generalization of any item that can be
+    # passed as arugement to a goal object
+    class GoalArg < BaseArg
     end # class
   end # module
 end # module